@@ -8,10 +8,10 @@ dynamic = ["version"]
 description = "The lightning-fast ASGI server."
 readme = "README.md"
 license = "BSD-3-Clause"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 authors = [
     { name = "Tom Christie", email = "tom@tomchristie.com" },
-    { name = "Marcelo Trylesinski", email = "marcelotryle@gmail.com" }
+    { name = "Marcelo Trylesinski", email = "marcelotryle@gmail.com" },
 ]
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -20,7 +20,6 @@ classifiers = [
     "License :: OSI Approved :: BSD License",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -38,8 +37,8 @@ dependencies = [
 
 [project.optional-dependencies]
 standard = [
-    "colorama>=0.4;sys_platform == 'win32'",
-    "httptools>=0.5.0",
+    "colorama>=0.4; sys_platform == 'win32'",
+    "httptools>=0.6.3",
     "python-dotenv>=0.13",
     "PyYAML>=5.1",
     "uvloop>=0.14.0,!=0.15.0,!=0.15.1; sys_platform != 'win32' and (sys_platform != 'cygwin' and platform_python_implementation != 'PyPy')",
@@ -51,7 +50,7 @@ standard = [
 uvicorn = "uvicorn.main:main"
 
 [project.urls]
-Changelog = "https://github.com/encode/uvicorn/blob/master/CHANGELOG.md"
+Changelog = "https://www.uvicorn.org/release-notes"
 Funding = "https://github.com/sponsors/encode"
 Homepage = "https://www.uvicorn.org/"
 Source = "https://github.com/encode/uvicorn"
@@ -60,11 +59,7 @@ Source = "https://github.com/encode/uvicorn"
 path = "uvicorn/__init__.py"
 
 [tool.hatch.build.targets.sdist]
-include = [
-    "/uvicorn",
-    "/tests",
-    "/requirements.txt",
-]
+include = ["/uvicorn", "/tests", "/requirements.txt"]
 
 [tool.ruff]
 line-length = 120
@@ -94,10 +89,9 @@ addopts = "-rxXs --strict-config --strict-markers"
 xfail_strict = true
 filterwarnings = [
     "error",
-    'ignore: \"watchgod\" is deprecated\, you should switch to watchfiles \(`pip install watchfiles`\)\.:DeprecationWarning',
     "ignore:Uvicorn's native WSGI implementation is deprecated.*:DeprecationWarning",
     "ignore: 'cgi' is deprecated and slated for removal in Python 3.13:DeprecationWarning",
-    "ignore: remove second argument of ws_handler:DeprecationWarning:websockets"
+    "ignore: remove second argument of ws_handler:DeprecationWarning:websockets",
 ]
 
 [tool.coverage.run]
@@ -132,8 +126,6 @@ py-win32 = "sys_platform == 'win32'"
 py-not-win32 = "sys_platform != 'win32'"
 py-linux = "sys_platform == 'linux'"
 py-darwin = "sys_platform == 'darwin'"
-py-gte-38 = "sys_version_info >= (3, 8)"
-py-lt-38 = "sys_version_info < (3, 8)"
 py-gte-39 = "sys_version_info >= (3, 9)"
 py-lt-39 = "sys_version_info < (3, 9)"
 py-gte-310 = "sys_version_info >= (3, 10)"
