@@ -5,7 +5,9 @@ import contextlib
 import logging
 import signal
 import sys
-from typing import Callable, ContextManager, Generator
+from collections.abc import Generator
+from contextlib import AbstractContextManager
+from typing import Callable
 
 import httpx
 import pytest
@@ -62,7 +64,7 @@ else:  # pragma: py-win32
 @pytest.mark.parametrize("exception_signal", signals)
 @pytest.mark.parametrize("capture_signal", signal_captures)
 async def test_server_interrupt(
-    exception_signal: signal.Signals, capture_signal: Callable[[signal.Signals], ContextManager[None]]
+    exception_signal: signal.Signals, capture_signal: Callable[[signal.Signals], AbstractContextManager[None]]
 ):  # pragma: py-win32
     """Test interrupting a Server that is run explicitly inside asyncio"""
 
