@@ -1,5 +1,5 @@
 from uvicorn.config import Config
 from uvicorn.main import Server, main, run
 
-__version__ = "0.32.0"
+__version__ = "0.34.0"
 __all__ = ["main", "run", "Config", "Server"]
