@@ -10,9 +10,10 @@ import socket
 import sys
 import threading
 import time
+from collections.abc import Generator, Sequence
 from email.utils import formatdate
 from types import FrameType
-from typing import TYPE_CHECKING, Generator, Sequence, Union
+from typing import TYPE_CHECKING, Union
 
 import click
 
@@ -82,8 +83,7 @@ class Server:
         logger.info(message, process_id, extra={"color_message": color_message})
 
         await self.startup(sockets=sockets)
-        if self.should_exit:
-            return
+        # FIX: make sure always execute the shutdown logic, even if server received a signal during startup
         await self.main_loop()
         await self.shutdown(sockets=sockets)
 
@@ -262,8 +262,9 @@ class Server:
         logger.info("Shutting down")
 
         # Stop accepting new connections.
-        for server in self.servers:
-            server.close()
+        if hasattr(self, "servers") and self.servers:
+            for server in self.servers:
+                server.close()
         for sock in sockets or []:
             sock.close()  # pragma: full coverage
 
@@ -284,10 +285,7 @@ class Server:
                 len(self.server_state.tasks),
             )
             for t in self.server_state.tasks:
-                if sys.version_info < (3, 9):  # pragma: py-gte-39
-                    t.cancel()
-                else:  # pragma: py-lt-39
-                    t.cancel(msg="Task cancelled, timeout graceful shutdown exceeded")
+                t.cancel(msg="Task cancelled, timeout graceful shutdown exceeded")
 
         # Send the lifespan shutdown event, and wait for application shutdown.
         if not self.force_exit:
@@ -308,8 +306,9 @@ class Server:
             while self.server_state.tasks and not self.force_exit:
                 await asyncio.sleep(0.1)
 
-        for server in self.servers:
-            await server.wait_closed()
+        if hasattr(self, "servers") and self.servers:
+            for server in self.servers:
+                await server.wait_closed()
 
     @contextlib.contextmanager
     def capture_signals(self) -> Generator[None, None, None]:
