@@ -9,15 +9,8 @@ if TYPE_CHECKING:
     ChangeReload: type[BaseReload]
 else:
     try:
-        from uvicorn.supervisors.watchfilesreload import (
-            WatchFilesReload as ChangeReload,
-        )
+        from uvicorn.supervisors.watchfilesreload import WatchFilesReload as ChangeReload
     except ImportError:  # pragma: no cover
-        try:
-            from uvicorn.supervisors.watchgodreload import (
-                WatchGodReload as ChangeReload,
-            )
-        except ImportError:
-            from uvicorn.supervisors.statreload import StatReload as ChangeReload
+        from uvicorn.supervisors.statreload import StatReload as ChangeReload
 
 __all__ = ["Multiprocess", "ChangeReload"]
