@@ -5,10 +5,11 @@ import os
 import signal
 import sys
 import threading
+from collections.abc import Iterator
 from pathlib import Path
 from socket import socket
 from types import FrameType
-from typing import Callable, Iterator
+from typing import Callable
 
 import click
 
