@@ -11,7 +11,7 @@ from gunicorn.arbiter import Arbiter
 from gunicorn.workers.base import Worker
 
 from uvicorn.config import Config
-from uvicorn.main import Server
+from uvicorn.server import Server
 
 warnings.warn(
     "The `uvicorn.workers` module is deprecated. Please use `uvicorn-worker` package instead.\n"
